Initial setup for flask app
I've created a basic flask app. It is not all implemented, currently I'm working
on the code to process the raw strings and get the requested unit ranges. I've
included some basic tests as well. This should represent much of the overall
structure of the app, but doesn't have database connections and the like yet.