Attempt to generalize seqvar
This is a more general function which should be able to compute the sequence
variation for any of the ranges we allow. I'm making a whole bunch of assumptions
about the naming scheme for the procedures, which I hope are right. Also, the
procedure returns two values but I'm only returning the first at the moment.
This is not ideal as the second is much more informative, however, we should
decide which is best to use in general. It looks like the second is more
informative so I would suggest that if performance is not an issue.