Begin to actually query the DB for variations
This integrates begins to integrate the code in utils.py with db.py. I'm aiming
to use the generic querying function with the test data stored in utils.py. This
will serve as a good test for the entire server. I have also changed the input
of the main function to get sequence variation to take only a dictionary. This
is because we do not yet do anything with sequence level information and may not
do anything in the near future. This way I remove it while letting things be
fairly flexible for future work.