Add templates and webcode to render them
We now have multiple possible output content types, application/json and
text/html. This is useful for both input and output. In the case of input we
will provide a simple page with a useful form to fill out. For output we can
produce a simple HTML page to display the data. Also, by allowing a JSON output
we can integrate this tool with other web sites and services.

Right now all templates are just stubs but they will be filled out later as we
settle on a design.