Use new database procedure
This commit is meant to change the current generic function to work with the new
more generic database procedure. To do this we need to always give 5 arguments
and make sure the arguments are null, or None in python, when there are no
matches. This commit adds that logic as well as changing the name to match the
new procedure name. However, we do not error checking on the inputs.