More realistic results display
I think that I had been over-complicating this -- I don't yet need child
templates to do what I want.  Includes a first attempt to repeat the
header periodically (every 20 rows, as a start; that can change, or
possibly become a user-selectable option, as long as we put a reasonable
default on it).