Refactor app and utils
I've refactored app and utils so allow for cleaner separation between parsing
inputs and computing variations. This means the parts of the site which compute
the variations are only called in the top level app module. We now separate out
parsing into ranges from getting variations. I have added some tests to make
sure the main parser in utils does what is meant to. I have improved unit id
parsing by using the parser in rnastructure to get the needed parts.

In addition, I added a production application. This app differs in that I've
attached a logger for easier debugging.