More detailed bad request exceptions
This should allow users a better idea of what is wrong with the request. We
return some more detailed information about what went wrong at any part of the
parsing of requests. We do not return information about allowed structures if
the user requests something we do not know about however. This should fixed.