Add more error checking for ranges and collections
Ranges and collections should be checked to make sure all required parts exist.
That means a range must have both a start and endpoint. In a collection, there
can be no empty parts to the collection. I've added some tests for these cases
as well as code to make these tests.