Refactor to use translations
I've implemented translating into the python code. This entailed a large change
to the type of data returned by the parsing code. We no longer return just
tuples we now give back the whole dictionary of the parsed unit id. This changes
lots of parts of the code.

In addition, I made the number of ranges limit a constant as well as cleaning up
some test code.