Fix errors caused by move to dictionaries
In moving to represent the parts of a collection as dictionaries I caused some
errors. The defaults for a missing entry were set to False. This causes problems
when we try to treat it as a dictionary. So I fix this by making the defaults an
empty dictionary instead of False. We also use .get to access parts of the
dictionary with the default set to the correct indication of missing. For a
chain that is '' to mean null in the query and for the numbers that is False.