Move from flask-mime to mimerender
mimerender does not seem to cause us to fail to produce static files. This is
very important as it makes developing the site difficult otherwise. mimerender
is not perfect in that it will serve up the default mimetype if the requested
one is not found. This is not ideal as it would be better to error out and
respond with the known mimetypes.