Partially complete form
The form can now validate and create unit ids. It will display error messages if
given invalid position input. It is also fairly forgiving about what valid input
is. We can skip ranges, only enter one endpoint to indicate a single unit and
such. However, this is not yet complete for several issues:

1. The alerts display will build up the message.
2. We do not yet build the chain select box.
3. We don't actually submit anything.

Still this is a good stopping point for the day.