Nearly complete form
This form now builds up and allows the selection of more than one alignment
(though we don't have such data yet) as well as generating the chain information
needed for the select boxes. We may be able to have a cleaner interface in cases
with only 1 chain, but that is not important yet. I have also allowed the chain
and alignment selectors to be auto selected if there is only 1 chain/alignment.
This will save the user some clicking.

In addition, I've fixed the issue with the previous commit where the error
messages would build up. Now they no do not build up.