Validate range size
We cannot process too large ranges, this checks user input for such things. The
range maximum size is hardcoded to 50 in the javascript, this is not ideal as it
is also hardcoded in the python code. It would be better to have this set by
python, but it works for now.