Fix bug where chains were not assigned to ranges
Not all ranges where getting a chain added. This was because the data-range
property of the chain option did not reflect the correct range index. I have
fixed this by looking at the parent data-range attribute if one does not exist
on the selected object. This one will always reflect the correct index.