Attempt to fix bug with 1S72 chains
1S72 has the chain we care about as chain 0. This causes problems because jQuery
is kind enough to convert the result of doing .data() to a number for us. I was
expecting a string that I could use .split() on. Since there is no .split() on
numbers this causes a failure. I'm attempting to fix this by converting to a
String and then doing .split().