Refactor so we always return the same data
The previous code would not always return the same data. If we posted we would
get a different structure back than doing a get. This is bad so I've moved the
logic to get the result into on function which is called when we need a result
from either the get or post methods.