Move landing jumbotron to the form page only
I think it looks better to only have the landing jumbotron on the form page. It
does not need to be on the result page. This will also make the form and result
page look very different, which will help people know when it has finished
loading. In addition, I've added an alert that shows how to use the form as
suggested by Neocles. This should make the site easier to use.