Add optional logging for queries used
This can be useful in several scenarios. For example we want to know how this
server is used in general so this helps. Plus we know what sorts of ranges
people are interested in trying. I'm making it optional because it is
technically possible using other logging. That said this makes it easier to
find.