Rework the design of the query page
The previous page did not have the form visible when the page loaded. This is
not ideal. I've reworked it to look more like JAR3D which I think has a nice
design. It moves the form to the top so people can easily find it but still
keeps all the help information available.