Catch any range size issues
Sometimes when processing a range in the database, we run into problems where
the range includes a large insertion. This means that we run into the recursion
limit on the database. Now we will catch these problems and then throw a
processing exception. This can be caught later and we can generate a more useful
error page.