Add a handler for invalid ranges
We now display more pleasant error pages if we fail to process a large range.
These large ranges should not create a nice page. Here we add an error handler
which will catch those issues and return a nice page. This requires the addition
of another template to display these pages.