Add a script to generate an options json file
We spend way too long generating the options to display every time we have a new
request. This script is the first step to getting around that by generating a
JSON file that we will load each time. This will speed up the process greatly,
while allow us to deploy new structures without having to redeploy the code.