Rework app to load the options/refactor examples
We now load the option information from a json file for each request. This
should speed up generating the main page considerably. This is the first step
along the pathway of speeding up the server. We can now turn to accessing the
rcad database happen in the background.

I've also pulled all example code out of app.py and put it into it's own module.
It is likely that examples and background processing will merge into the same
code. This is a step toward that. It also clears up app.py.