Reorganize all code
The code is about to get much more complex with a queue and background
processing code added. This means we will need to organize things a bit. As an
example, we are adding some code to deal with connecting to beanstalk, code for
connecting to rna3dhub databases, and possible some code a cache. This just
moves stuff around into the newer setup. I haven't fixed any imports yet so this
will break everything.