Update options.py script to new setup
The new setup requires different imports. In addition, the config file now
specifies the location of the options file. This means we no longer need to pass
it to main. I've also added an option to the command line to change the config
file to use, with the option defaulting to conf/config.json, the standard config
file.