Attempt at a version with background workers
Here I attempt to add the code that is needed for a version of this with
background workers and a cache. This requires several changes.

First the background workers need some sort of queue to log work to. For this
I'm using beanstalk which does this in a very simple and clean manor. Then we
need some place to temporarily store the results. For this we are using a
'cache'. The cache is really a database with a cache table. I'm planning on
using sqlite for this. This is not ideal as we should use something like
memcached, but this would add another moving part to the already complex
workflow. Plus that means more to install and configure on the move to the new
machine. Finally, we modify the queue code to pull stuff from the cache as
needed.

Since a job can be in more than one state we now need some sort of way of
telling that to users. This means we need templates for each state. I add new
templates and code to generate them as well. The changes in possible state also
alter the API, which needs to be documented.

Finally, I've stubbed out places to return the selected region as an alignment.
This should be possible now since things are cached and won't need to be
recomputed each time the same region is requested.