Add translation and propagate query info
We must translate the ranges as needed before we get the data for them. This
adds that logic into the worker. In addition, we must propagate information such
as the query id, the ranges, the units and such to the result. This is used
later on in generating pages so we have to do it here.