Correct logic for updating vs insert
So previously the logic was incorrect where we would detect if we should update
and then always use an insert. That's not a good idea as the process will always
fail in that case. I've corrected that issue. In addition, we did not specify
what object to update before. This corrects that issue so we do updates with a
where clause.