Add some code for generating alignment downloads
This can mostly generate strings for alignments. It fails with stockholm format
right now because we are using the AccssionID as the id of the sequence, but
there are several different sequences with the same AccessionID in some cases.
We may just need to use crw ids instead. I will ask Jamie, what the correct
thing to do here is.