Allow alignment export and correct json output
The json output was not being generated correctly before. This corrects that
issue. It was caused by a change in what was passed to each render. Previously
template as an argument followed by a dictionary, now it is just a dictionary with
a template key.

I have also registered alignment output as a possible mime type and provided it
as a possible output. It is now possible to ask for an alignment as a result of
a query. This is not yet visible to users but will be shortly. Also, someone
needs to code up a proper response if we are asked for data that has not yet
been processed.