Rework the cache and rendering
Rending has enough options that I like it to be in one module. I added the
render package to handle this. I may also want to move the stuff from alignment
in there later. But I need help with getting the alignment output correct.

I've also changed to use a real cache. That will be redis. We want this because
it can work as a cache with an expiration time on entries. In addition it
doesn't suffer from the 1MB limit on objects that memcahed does. As I don't yet
have a feeling how bit any entry will get it is better to be safe instead of
sorry.