Update fab file
We can now do a fab deploy to deploy any code changes. This will restart all
workers and create the new options file. A simple 'fab deploy' will do this.
Note that there is a possible issue where jobs can be buried forever if things
are restarted while running. Not good but I'm not sure how to deal with it yet.

I've also added an update_options task so we can just update the options file.