Create fishertest.r
Opens the csv file in filename (which should have no row or column names), strips any rows or columns with all zero entries, and runs an independence test.  Will do a chi squared test when appropriate, and uses a Monte Carlo estimate of Fisher's exact test with 10,000 replicates when chi squared is not appropriate (less than 80% of matrix entries have counts of 5 or greater).