Fix some issue noticed by Dr. Zirbel
He pointed out that the viewer sometimes shows weird triangle things. This seems
to be caused by having hydrogens in the list of atoms to display. Filter the
atoms to remove such things fixes the issue. This should be reported to the
author. Also, I've made the label text bold and styled it like a unit id.