Use a unique id for each sequence
The ids were not unique because I was not using the correct id, CRW id. This has
been fixed so now we have unique ids for all fragments in the produced
alignments. This fixes a bug where clustal files were not readable by clustal.
In addition, I may be able to create stockholm format files as well.