Use Description for alignments instead of AID Changes to three files support the use of "Description" for alignments, instead of the cryptic AID (which is also non-informative in isolation).
app.py contains most of the changes.  We can probably find a better way
to isolate the description from the options, but this works, and doesn't
produce a large performance hit in my testing.

worker.py and base.html needed only minor changes to support the new
piece of information to be displayed.