Add jar3d button for sequences
The idea will be to click on this button to send all the unique sequences to
jar3d. This button will be wired to some js that processes each unique sequence
to make sure it is valid for jar3d, not all gaps, has closing pairs, etc. It
will then submit it to jar3d and redirect as needed.