Possibly working version of jar3d integration
This now does a reasonable attempt at filtering sequences to pull out only the
ones which could be a valid sequence and then constructs a request for jar3d. It
will send things to the same host, so rnatest or rnaprod should be fine. This is
now ready for some testing on the test servers before moving it into production.