Attempt to fix production downloading issue
In production we cannot download alignment files, because the generated url uses
http not https. This is easily confirmed by opening up a results page and trying
to download. The links will all lead to http://, not https://, While the cause
of this is unclear a simple fix is to replace http with https as needed. This
makes this possible through a configuration flag. It is a flag since local
development deploys are unlikely to be https and we want to keep those working.